Refactors by extracting methods to new class
